+++ CHANGELOG.md
@@ -30,6 +30,7 @@ high state of flux, you're at risk of it changing without notice.
     - make `member` data-last (@gcanti)
     - make `elem` data-last (@gcanti)
     - make `lookup` data-last (@gcanti)
+    - make `lookupWithKey` data-last (@gcanti)
   - `ReadonlyArray`
     - export `unfold` (@gcanti)
     - make `lookup` data-last (@gcanti)
@@ -43,6 +44,7 @@ high state of flux, you're at risk of it changing without notice.
     - make `member` data-last (@gcanti)
     - make `elem` data-last (@gcanti)
     - make `lookup` data-last (@gcanti)
+    - make `lookupWithKey` data-last (@gcanti)
   - `Semigroup`
     - make `fold` data-last (@gcanti)